pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cowstudyapp"
version = "0.1.0"
authors = [
    { name = "Thomas Lipinski", email = "thomaslipinski@montana.edu" },
]
description = "A package for processing sensor data in pursuit of Precision Livestock Management."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.23.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "pytz>=2023.3",
    "pyproj>=3.7.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/cowstudyapp"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3